* 准备支持eventLoop复用，不知道能不能写出来，不知道有没有效果

+++ baseio-codec/src/main/java/com/generallycloud/baseio/codec/protobase/ProtobaseFutureImpl.java
@@ -260,7 +260,7 @@ public class ProtobaseFutureImpl extends AbstractChannelFuture implements Protob
         if (newcount > binaryWriteBuffer.length) {
             binaryWriteBuffer = Arrays.copyOf(binaryWriteBuffer, binaryWriteBuffer.length << 1);
         }
-        binaryWriteBuffer[binaryWriteSize] = (byte) b;
+        binaryWriteBuffer[binaryWriteSize] = b;
         binaryWriteSize = newcount;
     }
